Add the 'sealedunions' prefix
