+++ conjure-java-core/src/test/resources/sealedunions/sealedunions/com/palantir/product/StringExample.java
@@ -1,4 +1,4 @@
-package com.palantir.product;
+package sealedunions.com.palantir.product;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonSetter;
