+++ conjure-java-core/src/test/resources/sealedunions/sealedunions/com/palantir/product/Union.java
@@ -1,4 +1,4 @@
-package com.palantir.product;
+package sealedunions.com.palantir.product;
 
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
@@ -120,11 +120,11 @@ public sealed interface Union {
         @Override
         public boolean equals(Object other) {
             return this == other
-                    || (other instanceof com.palantir.product.FooWrapper
-                            && equalTo((com.palantir.product.FooWrapper) other));
+                    || (other instanceof sealedunions.com.palantir.product.FooWrapper
+                            && equalTo((sealedunions.com.palantir.product.FooWrapper) other));
         }
 
-        private boolean equalTo(com.palantir.product.FooWrapper other) {
+        private boolean equalTo(sealedunions.com.palantir.product.FooWrapper other) {
             return this.value.equals(other.value);
         }
 
@@ -168,11 +168,11 @@ public sealed interface Union {
         @Override
         public boolean equals(Object other) {
             return this == other
-                    || (other instanceof com.palantir.product.BarWrapper
-                            && equalTo((com.palantir.product.BarWrapper) other));
+                    || (other instanceof sealedunions.com.palantir.product.BarWrapper
+                            && equalTo((sealedunions.com.palantir.product.BarWrapper) other));
         }
 
-        private boolean equalTo(com.palantir.product.BarWrapper other) {
+        private boolean equalTo(sealedunions.com.palantir.product.BarWrapper other) {
             return this.value == other.value;
         }
 
@@ -216,11 +216,11 @@ public sealed interface Union {
         @Override
         public boolean equals(Object other) {
             return this == other
-                    || (other instanceof com.palantir.product.BazWrapper
-                            && equalTo((com.palantir.product.BazWrapper) other));
+                    || (other instanceof sealedunions.com.palantir.product.BazWrapper
+                            && equalTo((sealedunions.com.palantir.product.BazWrapper) other));
         }
 
-        private boolean equalTo(com.palantir.product.BazWrapper other) {
+        private boolean equalTo(sealedunions.com.palantir.product.BazWrapper other) {
             return this.value == other.value;
         }
 
@@ -275,11 +275,11 @@ public sealed interface Union {
         @Override
         public boolean equals(Object other) {
             return this == other
-                    || (other instanceof com.palantir.product.UnknownWrapper
-                            && equalTo((com.palantir.product.UnknownWrapper) other));
+                    || (other instanceof sealedunions.com.palantir.product.UnknownWrapper
+                            && equalTo((sealedunions.com.palantir.product.UnknownWrapper) other));
         }
 
-        private boolean equalTo(com.palantir.product.UnknownWrapper other) {
+        private boolean equalTo(sealedunions.com.palantir.product.UnknownWrapper other) {
             return this.type.equals(other.type) && this.value.equals(other.value);
         }
